pyproject.toml
[project]
name = "socialomni-benchmark"
version = "0.1.0"
description = "A benchmark framework for evaluating native joint reasoning in omni-modal models"
readme = "README.md"
requires-python =  ">=3.10,<3.11"
dependencies = [
    "torch==2.7.1",
    "torchvision==0.22.1",
    "torchaudio==2.7.1",
    "transformers==4.57.1",
    "accelerate==0.34.2",
    "triton==3.3.1",
    "xformers>=0.0.29.post3",
    "timm>=1.0.20",
    "einops>=0.8.1",
    "qwen-omni-utils>=0.0.8",
    "vllm==0.10.0",
    "audioread>=3.0.1",
    "av>=15.0.0",
    "librosa>=0.11.0",
    "soundfile>=0.13.1",
    "opencv-python>=4.12.0.88",
    "numpy>=2,<2.3.0",
    "moviepy>=2.2.1",
    "decord>=0.6.0",
    "pandas>=2.3.1",
    "fastparquet>=2024.11.0",
    "openpyxl>=3.1.5",
    "pyyaml>=6.0.2",
    "rarfile>=4.2",
    "py7zr>=1.0.0",
    "git-filter-repo>=2.47.0",
    "ipython>=8.37.0",
    "flask>=3.1.1",
    "openai>=1.87.0,<=1.90.0",
    "setuptools>=77.0.3,<80",
    "flash-attn>=2.8.3",
    "python-dotenv",
    "matplotlib>=3.10.7",
    "easydict>=1.13",
    "diffusers>=0.35.2",
    "omegaconf>=2.3.0",
    "fire>=0.7.1",
    "deepspeed>=0.18.0",
    "ujson>=5.11.0",
    "lightning>=2.5.5",
    "snac>=1.2.1",
    "openai-whisper>=20250625",
    "kaldiio>=2.18.1",
    "beartype>=0.22.3",
    "pypdf2>=3.0.1",
    "google-genai>=1.49.0",
    "httpx[socks]>=0.28.1",
    "s2wrapper",
    "whisper>=1.1.10",
    "clip",
    "imagesize>=1.4.1",
    "cairosvg>=2.8.2",
]

[[tool.uv.index]]
name = "tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true

[tool.uv.sources]
transformers = { git = "https://github.com/huggingface/transformers" }
s2wrapper = { git = "https://github.com/bfshi/scaling_on_scales.git" }
clip = { git = "https://github.com/openai/CLIP.git" }

[tool.ruff]
line-length = 88
target-version = "py310"
select = ["E", "F", "W", "I", "B", "C4", "UP"]
ignore = ["E501", "B008"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
